move debugLogger to Utils
